feat(v46+v47): code consolidation + documentation restructure
v46 code changes:
- 212 source files, 85,116 lines, 2,075 unit tests (0 failures)
- 44 integration test files
- Retrieval fusion engine, cognitive loop, intent executor, KG builder
- New benchmarks/ framework (Python, uv) replacing old bench/

v47 documentation restructure:
- CLAUDE.md: 589→278 lines, split into .claude/rules/ (3 files)
- AGENTS.md: 462→126 lines, ariadne-thread 9-section standard
- README.md/README_zh.md: updated to v46 stats
- docs/milestones/: template + INDEX + v34/v35/v41/v43/v46 summaries
- Deleted 50+ outdated docs (design-node*, audit*, benchmark reports)
- Deleted 28 transient .log files from root
- MEMORY.md: cleaned, historical data archived
- benchmarks/README.md: comprehensive framework docs

Quality gates passed:
- cargo build: OK
- cargo test --lib: 2,075 passed, 0 failed
- cargo clippy -D warnings: 0 errors

+++ tests/v11_metrics.rs
@@ -500,7 +500,7 @@ fn test_end_to_end_scenario_with_metrics() {
     // Verify targets
     assert!(time_reduction >= TIME_REDUCTION_TARGET_PCT || elapsed_ms < baseline_time,
         "Time should be reduced or be less than baseline");
-    assert!(tool_reduction >= TOOL_REDUCTION_TARGET_PCT || tool_calls < baseline_tools as usize,
+    assert!(tool_reduction >= TOOL_REDUCTION_TARGET_PCT || tool_calls < baseline_tools,
         "Tool calls should be reduced or be less than baseline");
     assert!(token_pct <= TOKEN_TARGET_PCT * 1.5, // Allow some slack for stub
         "Token percentage should be reasonably close to target");
